pyproject.toml
[tool.poetry]
name = "fastapi-python"
version = "0.1.0"
description = ""
authors = ["Dmitry Morozov <morozov124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
fastapi = {extras = ["all"], version = "^0.115.12"}
mypy = "^1.15.0"
aiofiles = "^24.1.0"
itsdangerous = "^2.2.0"
bcrypt = "^4.3.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
cryptography = "^44.0.3"
slowapi = "^0.1.9"
fastapi-limiter = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
